requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.8.5
asyncio==3.4.3
